Flash messages

+++ delete.php
@@ -11,7 +11,7 @@ if (!isset($_SESSION["user"])) {
 
 $id = $_GET["id"];
 
-$statement = $conn->prepare("SELECT * FROM contacts WHERE id = :id");
+$statement = $conn->prepare("SELECT * FROM contacts WHERE id = :id LIMIT 1");
 $statement->execute([":id" => $id]);
 
 if ($statement->rowCount() == 0) {
@@ -20,6 +20,16 @@ if ($statement->rowCount() == 0) {
   return;
 }
 
+$contact = $statement->fetch(PDO::FETCH_ASSOC);
+
+if ($contact["user_id"] !== $_SESSION["user"]["id"]) {
+  http_response_code(403);
+  echo("HTTP 403 UNAUTHORIZED");
+  return;
+}
+
 $conn->prepare("DELETE FROM contacts WHERE id = :id")->execute([":id" => $id]);
+
+$_SESSION["flash"] = ["message" => "Contact {$contact['name']} deleted."];
 
 header("Location: home.php");
